+++ src/index.js
@@ -2,16 +2,14 @@ const _ = require('lodash');
 const multer = require('koa-multer');
 const path = require('path');
 const fs = require('fs');
-var crypto = require('crypto')
-
-//将上传的信息保存在内存中
-var datas = {}
+const crypto = require('crypto');
 
 module.exports = {
   bind: (fpm) => {
-    const config = fpm.getConfig('upload', {
-      dir: 'public/uploads',
+    const config = Object.assign({
+      dir: 'public/uploads/',
       field: 'upload',
+      uploadRouter: '/upload',
       base: '/uploads/',
       accept: [
         'application/octet-stream',
@@ -19,10 +17,10 @@ module.exports = {
         'application/zip',
         'application/x-zip-compressed',
         'image/png',
-        'image/jpeg'],    // 上传的文件类型限制
+        'image/jpeg'],    // Allowed type
       limit: 5,  // 5MB
 
-    })
+    }, fpm.getConfig('upload', {}));
     const fileFilter = (req, file, cb) =>{
       if(_.indexOf(config.accept, file.mimetype)> -1){
           cb(null, 1)
@@ -32,39 +30,42 @@ module.exports = {
     }
 
     const dest = path.join(fpm.get('CWD'), config.dir);
-    var storage = multer.diskStorage({
-      destination: function (req, file, cb) {
+    const storage = multer.diskStorage({
+      destination: (req, file, cb) =>{
         cb(null, dest)
       },
-      filename: function (req, file, cb) {
-        crypto.pseudoRandomBytes(16, function (err, raw) {
+      filename: (req, file, cb) =>{
+        crypto.randomBytes(16, (err, raw) =>{
           cb(err, err ? undefined : (raw.toString('hex') + path.extname(file.originalname)))
         })
       }
     })
+
+    // define the default upload function
     const upload = multer({ fileFilter: fileFilter, storage: storage, limits: { fileSize: config.limit * 1024 * 1024 }});
 
-    // 上传表单以file为文件的字段
+    // default form handler
     const defaultHandler = upload.single(config.field);
+
 
     // 捕获异常
     const handler = async (ctx, next) => {
         try{
             await defaultHandler(ctx, next)
-            let data = ctx.req.file;
-            data.hash = path.basename(data.filename, path.extname(data.filename));
-            datas[data.hash] = data;
+            const data = ctx.req.file;
+            const { filename } = data;
+            data.hash = path.basename(filename, path.extname(filename));
+
             fpm.publish('#upload/success', data);
-            datas['latest'] = data;
+
             ctx.body = {
                 errno: 0,
                 uploaded: true,
-                url: config.base + data.filename,
-                data: _.assign({
-                    id: data.hash,
-                    path: path.join(config.dir, data.filename),
-                    url: '/download/' + data.hash,
-                })
+                url: config.base + filename,
+                data: {
+                  hash: data.hash,
+                  path: config.base + filename,
+                }
             }
         }catch(e){
             ctx.body = {
@@ -75,18 +76,10 @@ module.exports = {
         }
     }
     fpm.registerAction('FPM_ROUTER', () => {
-      console.log('Run BEFORE_SERVER_START Actions')
       const router = fpm.createRouter();
-      router.post('/upload', handler);
 
-      // 下载文件的路由
-      router.get('/download/:id', async (ctx, next) => {
-          let data = datas[ctx.params.id]
-          ctx.type = data.mimetype
-          ctx.attachment(data.originalname)
-          ctx.body = await fs.createReadStream(data.path)
-          await next()
-      })
+      // bind the upload url handler.
+      router.post(config.uploadRouter, handler);
 
       fpm.bindRouter(router);
     })