+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+Remove the download router.
+
+Add uploadRouter in config.json
 ## 0.0.1
 
 Change `script` in `package.json`